Updated views for books
